gts_engine_service/requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.2.0
altair==4.2.0
anyio==3.6.2
async-timeout==4.0.2
attrs==22.1.0
backports.zoneinfo==0.2.1
blinker==1.5
cachetools==5.2.0
charset-normalizer==2.1.1
cheroot==8.6.0
click==8.1.3
commonmark==0.9.1
contourpy==1.0.5
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
fastapi==0.85.1
filelock==3.8.0
fonttools==4.38.0
frozenlist==1.3.1
fsspec==2022.10.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.29
google-auth==2.13.0
google-auth-oauthlib==0.4.6
grpcio==1.50.0
h11==0.14.0
huggingface-hub==0.10.1
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
jaraco.functools==3.5.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
kiwisolver==1.4.4
loralib==0.1.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
more-itertools==9.0.0
multidict==6.0.2
numpy==1.23.4
oauthlib==3.2.2
packaging==21.3
pandas==1.5.1
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
protobuf==3.20.3
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
pydeck==0.8.0b4
pyDeprecate==0.3.1
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytorch-lightning==1.5.10
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
rjieba==0.1.11
roformer==0.4.3
rsa==4.9
sacremoses==0.0.53
scikit-learn==1.1.2
scipy==1.9.3
semver==2.13.0
sh==1.14.3
six==1.16.0
sklearn==0.0
smmap==5.0.0
sniffio==1.3.0
sse-starlette==1.1.6
starlette==0.20.4
streamlit==1.13.0
streamlit-tags==1.2.8
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tokenizers==0.12.1
toml==0.10.2
toolz==0.12.0
torchmetrics==0.10.0
tornado==6.2
tqdm==4.64.1
transformers==4.18.0
typing_extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
uvicorn==0.19.0
validators==0.20.0
watchdog==2.1.9
web.py==0.62
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.9.0
